+++ src/main/java/gift/controller/restcontroller/ProductRestController.java
@@ -1,7 +1,7 @@
 package gift.controller.restcontroller;
 
-import gift.controller.dto.ProductRequest;
-import gift.controller.dto.ProductResponse;
+import gift.controller.dto.request.ProductRequest;
+import gift.controller.dto.response.ProductResponse;
 import gift.model.ProductDao;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Min;
